requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.3
onnxruntime==1.16.3
pydantic>=2.7.0
opencv-python==4.9.0.80
scipy==1.11.4
transformers==4.37.2

torchaudio==2.1.0
requests==2.31.0
tqdm==4.66.1
soundfile==0.12.1
pydub==0.25.1
Librosa==0.10.1

pandas==2.2.0
scikit-learn==1.4.0
torch==2.2.0
torchvision==0.17.0
onnx==1.15.0
skl2onnx==1.16.0
matplotlib==3.8.2
seaborn==0.13.1
pyarrow==12.0.1